requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.46
Flask-SQLAlchemy==3.1.1
fpdf2==2.7.0
python-dateutil==2.8.2
Werkzeug==3.0.1
wikipedia-api==0.9.0
requests==2.31.0
